@cmacknz 2023-06-15
You can use https://pkg.go.dev/golang.org/x/perf/cmd/benchstat to generate direct comparisons between the before and after benchmarks. Would make it easier to see the improvements.