@cmacknz 2023-06-15
@alexsapran has a performance framework that should be able to run these changes in an actual instance of Filebeat so we can see the improvement to the entire pipeline outside the of just the Go benchmarks here.

I would like to see these results before we merge this, both to sanity check the changes and to see what the total affect on ingestion performance is.